updating design of the footrpint

+++ Valhalla.Charting/StockChart.cs
@@ -121,8 +121,8 @@ namespace Valhalla.Charting
         public double Price { get; set; }
         public double Volume { get; set; }
         public long Delta { get; set; }
-        public long BuyVolume { get; set; }
-        public long SellVolume { get; set; }
+        public double BuyVolume { get; set; }
+        public double SellVolume { get; set; }
         public long Trades { get; set; }
         public long BuyTrades { get; set; }
         public long SellTrades { get; set; }
